Initial implementation of `TemplateNodeInfo`
