+++ cluster-autoscaler/cloudprovider/upcloud/pkg/github.com/upcloudltd/upcloud-go-api/v8/upcloud/client/error.go
@@ -1,0 +1,23 @@
+package client
+
+import "fmt"
+
+type ErrorType int
+
+const (
+	ErrorTypeError = iota
+	ErrorTypeProblem
+)
+
+// Error represents an error returned from the client. Errors are thrown when requests don't have a successful status code
+type Error struct {
+	ErrorCode    int
+	ErrorMessage string
+	ResponseBody []byte
+	Type         ErrorType
+}
+
+// Error implements the Error interface
+func (e *Error) Error() string {
+	return fmt.Sprintf("%d: %s", e.ErrorCode, e.ErrorMessage)
+}
