+++ cluster-autoscaler/cloudprovider/upcloud/pkg/github.com/upcloudltd/upcloud-go-api/v8/upcloud/service/stubs.go
@@ -1,0 +1,21 @@
+
+package service
+
+type Account interface{}
+type Firewall interface{}
+type Host interface{}
+type IPAddress interface{}
+type LoadBalancer interface{}
+type Tag interface{}
+type Storage interface{}
+type ObjectStorage interface{}
+type ManagedDatabaseServiceManager interface{}
+type ManagedDatabaseUserManager interface{}
+type ManagedDatabaseLogicalDatabaseManager interface{}
+type Permission interface{}
+type ServerGroup interface{}
+type Server interface{}
+type ManagedObjectStorage interface{}
+type Gateway interface{}
+type Partner interface{}
+
