+++ cluster-autoscaler/cloudprovider/upcloud/vendor.sh
@@ -31,15 +31,14 @@ sdk_download () {
     curl -sO --output-dir $1 $2
 }
 
-sdk_download $sdk_dir "${sdk_url}/upcloud/{kubernetes.go,problem.go,utils.go,label.go,ip_address.go,network.go}"
+sdk_download $sdk_dir "${sdk_url}/upcloud/{kubernetes.go,problem.go,utils.go,label.go,ip_address.go,network.go,plan.go,price.go,storage.go,timezone.go,zone.go}"
 sdk_download $sdk_dir/client "${sdk_url}/upcloud/client/{client,error}.go"
 sdk_download $sdk_dir/request "${sdk_url}/upcloud/request/{kubernetes.go,request.go,network.go}"
-sdk_download $sdk_dir/service "${sdk_url}/upcloud/service/{kubernetes.go,service.go,network.go}"
+sdk_download $sdk_dir/service "${sdk_url}/upcloud/service/{kubernetes.go,service.go,network.go,cloud.go,retry.go}"
 
 echo "
 package service
 
-type Cloud interface{}
 type Account interface{}
 type Firewall interface{}
 type Host interface{}
@@ -54,6 +53,9 @@ type ManagedDatabaseLogicalDatabaseManager interface{}
 type Permission interface{}
 type ServerGroup interface{}
 type Server interface{}
+type ManagedObjectStorage interface{}
+type Gateway interface{}
+type Partner interface{}
 " > $sdk_dir/service/stubs.go
 
 find $sdk_dir -name "*.go" -exec sed -i 's#"'${UPCLOUD_SDK_PACKAGE}'#"k8s.io/autoscaler/cluster-autoscaler/cloudprovider/upcloud/pkg/'${UPCLOUD_SDK_PACKAGE}'#gI' {} \;
